:recycle: refactor(viewer): split telegram clipboard formatter
